feat: 유저정보 조회 기능

+++ backend/src/main/java/com/yunyun/meallog/global/interceptor/JwtInterceptor.java
@@ -28,8 +28,8 @@ public class JwtInterceptor implements HandlerInterceptor {
             String token = header.substring(7);
 
             if (jwtUtil.isTokenValid(token)) {
-                // String -> Long 변환
-                Long userId = Long.parseLong(jwtUtil.getUserId(token));
+                
+                Integer userId = Integer.parseInt(jwtUtil.getUserId(token));
                 request.setAttribute("userId", userId);
             } else {
                 response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
